requirements.txt
Flask
numpy
tensorflow
matplotlib
fpdf
python-dotenv
google-generativeai
Werkzeug